fix(): refactor app.ts
